requirements/common.txt
# ^^^^^^^^^^^^^^^^^^^^^^^^
# DJANGO & EXTENSIONS
# Also requires: None
# ^^^^^^^^^^^^^^^^^^^^^^^^
Django==2.0.3
django-extensions==2.0.6
django-model-utils==3.1.1
django-shortuuidfield==0.1.3
djangorestframework==3.7.7

# ^^^^^^^^^^^^^^^^^^^^^^
# DATABASE MODULES
# ^^^^^^^^^^^^^^^^^^^^^^
psycopg2-binary==2.7.4

# ^^^^^^^^^^^^^^^^^^^^^^
# TESTING
# ^^^^^^^^^^^^^^^^^^^^^^
factory-boy==2.10.0

# ^^^^^^^^^^^^^^^^^^^^^^
# STANDALONE
# ^^^^^^^^^^^^^^^^^^^^^^
shortuuid==0.5.0
djangorestframework-jwt==1.11.0

# ^^^^^^^^^^^^^^^^^^^^^^
# INTEGRATIONS
# ^^^^^^^^^^^^^^^^^^^^^^

